feat(core): get more context in `IgnoredLints`
