+++ harper-core/src/fat_token.rs
@@ -4,7 +4,7 @@ use crate::TokenKind;
 
 /// A [`Token`](crate::Token) that holds its content as a fat [`Vec<char>`] rather than as a
 /// [`Span`](crate::Span).
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, PartialOrd)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, PartialOrd, Hash)]
 pub struct FatToken {
     pub content: Vec<char>,
     pub kind: TokenKind,
