+++ harper-core/src/ignored_lints/mod.rs
@@ -1,6 +1,9 @@
+mod lint_context;
+
 use std::hash::{DefaultHasher, Hash, Hasher};
 
 use hashbrown::HashSet;
+use lint_context::LintContext;
 use serde::{Deserialize, Serialize};
 
 use crate::{linting::Lint, Document};
@@ -22,26 +25,10 @@ impl IgnoredLints {
     }
 
     fn hash_lint_context(&self, lint: &Lint, document: &Document) -> u64 {
-        let problem_tokens = document.token_indices_intersecting(lint.span);
-        let prequel_tokens = lint
-            .span
-            .with_len(2)
-            .pulled_by(2)
-            .map(|v| document.token_indices_intersecting(v))
-            .unwrap_or_default();
-        let sequel_tokens = document.token_indices_intersecting(lint.span.with_len(2).pushed_by(2));
+        let context = LintContext::from_lint(lint, document);
 
         let mut hasher = DefaultHasher::default();
-
-        problem_tokens
-            .into_iter()
-            .chain(prequel_tokens)
-            .chain(sequel_tokens)
-            .flat_map(|idx| document.get_token(idx))
-            .for_each(|tok| tok.kind.hash(&mut hasher));
-
-        let lint_hash = lint.spanless_hash();
-        lint_hash.hash(&mut hasher);
+        context.hash(&mut hasher);
 
         hasher.finish()
     }
